Move core/package_run.go into runtime/parser.go
The package_run.go file was in 'core' package but it relied on
'runtime' package to be imported. Problem is that while implementing
some new feature, I need to access 'core' package from within the
'runtime' package and BUUUM, cyclic import problem.

So with this commit we move the file into 'runtime' package. It feels
more natural to have it here in this package since everything inside
it is runtime-related. Also, the logic inside this file is focused on
parsing meta/run.yaml file into internal structs, therefore we rename
it into parser.go (was package_run.go before).
